FileBot - Checks if the system is supported

+++ utils/filebot/filebot-uninstaller.sh
@@ -4,10 +4,15 @@
 source "$SCRIPTPATH/inc/commons.sh"
 source "$SCRIPTPATH/inc/header.sh"
 echo -e "${GREEN}AtoMiC $APPTITLE Uninstaller Script$ENDCOLOR"
-source "$SCRIPTPATH/inc/pause.sh"
-echo
-echo -e "$YELLOW--->Uninstalling snap $APPTITLE package...$ENDCOLOR"
-snap remove $APPNAME
-source "$SCRIPTPATH/inc/app-uninstall-confirmation.sh"
-source "$SCRIPTPATH/inc/thankyou.sh"
+if IsSystemdSupported; then
+    source "$SCRIPTPATH/inc/pause.sh"
+    echo
+    echo -e "$YELLOW--->Uninstalling snap $APPTITLE package...$ENDCOLOR"
+    snap remove $APPNAME
+    source "$SCRIPTPATH/inc/app-uninstall-confirmation.sh"
+    source "$SCRIPTPATH/inc/thankyou.sh"
+else
+    echo -e "${RED}Distro not curently supported." \
+            "Please consider upgrading to a newer version$ENDCOLOR"
+fi
 source "$SCRIPTPATH/inc/exit.sh"
